Separate out tests for ServerFormUseCases
These test file names are pretty long, but they allow IntelliJ
to navigate between the file and the test.
